Action sequence while matching
UPDATE: Action sequence doesn't need to be calculated, as they are
already stacked the right way. What is needed is only the number of
arguments for each Action, for using and popping them in a `pt_data` stack.
`pt_run_actions` is now iterative, doesn't allocate that much
memory and its complexity changed from O(n^2) to O(n) =]